organize classes
https://github.com/metasfresh/metasfresh/issues/4524
