+++ de.metas.inbound.mail/src/main/java/de/metas/inbound/mail/InboundEMailService.java
@@ -12,6 +12,7 @@ import org.springframework.stereotype.Service;
 
 import com.google.common.collect.ImmutableList;
 
+import de.metas.inbound.mail.config.InboundEMailConfig;
 import lombok.NonNull;
 
 /*
