+++ de.metas.inbound.mail/src/main/java/de/metas/inbound/mail/config/InboundEMailConfigId.java
@@ -1,6 +1,12 @@
-package de.metas.inbound.mail;
+package de.metas.inbound.mail.config;
 
-import java.util.Set;
+import java.util.UUID;
+
+import org.adempiere.util.Check;
+
+import lombok.EqualsAndHashCode;
+import lombok.NonNull;
+import lombok.ToString;
 
 /*
  * #%L
@@ -12,19 +18,43 @@ import java.util.Set;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
 
-public interface InboundEMailConfigChangedListener
+@EqualsAndHashCode
+@ToString
+public final class InboundEMailConfigId
 {
-	void onInboundEMailConfigChanged(Set<InboundEMailConfigId> changedConfigIds);
+	public static final InboundEMailConfigId ofRepoId(final int repoId)
+	{
+		Check.assumeGreaterThanZero(repoId, "repoId");
+		return new InboundEMailConfigId(String.valueOf(repoId));
+	}
+
+	public static final InboundEMailConfigId random()
+	{
+		return new InboundEMailConfigId(UUID.randomUUID().toString());
+	}
+
+	private final String id;
+
+	private InboundEMailConfigId(@NonNull final String id)
+	{
+		this.id = id;
+	}
+
+	public String getAsString()
+	{
+		return id;
+	}
+
 }
